requirements.txt
googlemaps
google-cloud-secret-manager
python-dotenv
google-genai
Flask
google-cloud-storage